Move shellcode imports into the functions that use them. (#124)
When a module is imported, the code at the top of the class is also
executed (because it's global to the class and not part of the class's
`__init__`). This is not the behavior I am expecting, so I am moving the
imports into the function that uses the shellcode functions instead.