Reread memory before writing. (#144)
We're doing a weird thing with clarity where we are infinitely scanning
memory for strings. When we find certain patterns, we overwrite them.

This is a problem when memory is always changing; memory is being
de-allocated and re-allocated at any given time. Something that was
found with `pattern_scan_all` may no longer be the case anymore, but we
could write over valid data that was allocated and cause crashes because
of it.

This isn't going to fix 100% of these problems, but it should
significantly reduce the chance of this happening by reading from the
address immediately before writing to it. If the data doesn't match, the
the previous read was stale and we throw it away altogether.