Fix pattern scan oom (#148)
Have been seeing an issue where `scan_pattern_page` has been reading
more bytes than it's supposed to, resulting in a
[MemoryError](https://docs.python.org/3/library/exceptions.html#MemoryError)
error.

This [SO post](https://stackoverflow.com/a/42057650) seems to suggest
that this is thrown if Python thinks it will OOM -- not that it has. In
any case, I've run into this a few times now and it's always during
pattern scanning, so we'll catch these and just continue on.