refactor: don't access raw json files and use db instead. (#177)
Currently, there is a mixture of reading json files and accessing the db
depending on what is being done. This refactors everything to read from
the db.

When we run an update, we'll keep the files in memory and write them to
the database instead of writing them to the `misc_files` directory.

This also cleans up how the `merge.xlsx` file is parsed and imported,
making it more resilient when forgetting to truncate the spreadsheet's
extra rows as well as assuming a `replace` if the data exists (instead
of running a `select`, checking it and then extra code to `update` or
`insert`).

Because I always seem to dog pile issues onto a single PR, this also
fixes several inefficiencies around how dict lookups are being done.
Instead of iterating through every value in a dict until we find
something, we just do a direct lookup instead.

More dog piling -- there is a performance issue with the party hook. As
it is triggered very frequently and the code that runs within the hook
takes too long to execute, it causes stuttering in busy areas (like
Megistris server 1). This hook has been removed and the old scan has
been reinstated. I'd rather a slower, continuous scan over game
stuttering.

Final add: Adds support to replace placeholder tags like `<pnplacehold>`
and `<snplacehold>` in the m00_strings table. This allows us to match
more strings in our custom files we capture.