feat: Add ability to scan only for data pages (#225)
pymem's `pattern_scan_all` currently scans the entirety of the process's
memory region using `scan_pattern_page` to iterate over each memory page
to evaluate if we want to read it in and run a query against it.

clarity generally only cares about memory that:

- is of type "PRIVATE" or "MAPPED" (we don't care about memory that is
already mapped to an image, which exists in a module)
- has anything but PAGE_READWRITE defined in its protections (if we're
scanning for data, the intent is to modify it)

this adds a `data_only` flag to `pattern_scan` so that we only scan for
readable, writable memory. This offers a slight boost (on my pc) of
about 200ms when running the main `run_scans` loop. This isn't
significant, but as we may scan several hundreds of times over the
course of a game session, this time adds up.