Update network_text.py (#281)
-C_ITMR_STITLE is the text that appears in the same spot as, for
example, the "Bonus Time Remaining" text but is server side text. It
doesn't appear often from what I can tell, but it does appear during the
Dolboard race and testing shows it doesn't affect performance.

-CAS_gambler is a player's name shown during casino games like sugoroku
messages such as "<name> lands on a Church space!" This is barely called
so it shouldn't affect performance. I tested a game myself and didn't
notice anything

-C_MERCENARY is the name that appears when hiring people from the tavern
in that UI and that related text, and the "Your contract with <name>
expires in X minutes" dracky message. Tested with it and doesn't seem to
affect much since it only shows up in these specific scenarios