Update network_text.py (#319)
Removes the following vars:

-B_ACTOR/B_TARGET: These are used in combat messages in both the
communication window and the bottom-middle screen battle messages. If
player/npc name scans are on, these vars are overwritten with their
scanned translated versions with no weird spaces. If either of these
scans are turned off, then they are overwritten by their translated
versions caught by the network hook, causing the weird spacing. If a
player has player/npc names turned off, these should be in Japanese
anyway.

L_GOODS: Used exclusively in the DQX Shop vendor in Megistris to print
the name of the item you are buying to dialogue (which is now breaking
due to the byte limit stuff.) This is mostly redundant as the player is
able to see the name of the item they are buying beforehand without the
network hook text.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>