fix/refactor: signal to main entrypoint when processes are spawned (#340)
This PR encompasses several items:

### main.py entrypoint

The entrypoint into `main.py` that spawned multiple processes for the
scans had a manual sleep added to wait for the processes to be ready.
This was obviously a bad workaround as this this doesn't accurately wait
to display "done!" to the user and could still be processing in the
backend before displaying the message. In the event a user logged into
their character after being told to log in and we were still hooking, we
could cause the game to crash.

This of course assumes the user is actually reading the console output
after running and waiting for it to say "done".

**change**: This now starts a process and sends an event handler to the
target function that spawns the process. The function then needs to
signal that it's ready back to the event handler for the script to
continue. There's a timeout period in case the process takes too long -
which would indicate a problem.

### remove dependency

**change**: Removes the `click` dependency and uses the standard
`argparse` library instead. This is made just to remove dependent
libraries - especially as it was used just in the entrypoint script.

### refactor scanning

**change**: Breaks the scans up into separate files based on how they're
called. As the scans in dqxclarity are spawned into individual processes
to be run in the background, they're now grouped based on how they're
called to make it easier to understand how they run. There's a
`manager.py` that still handles how the scans are executed.

### integrity scan check

**change**: Corrects the way we check if the integrity check address can
be found. In the event a user launches the program and the game patched,
or the user launched dqxclarity twice (after closing the first
instance), we'll realize this and send an error to the user to let them
know what happened.