port hash and blowfish logging to hooks. (#392)
This ports the great work that Andor did into python hooks since we can
stay hooked without the game crashing now.

As part of getting this to work, I also need to back up the register
flags as well. Instead of needing to read from the remote state that we
are storing the registers, we'll just use asm to pop them back onto the
stack.