fix: player pets getting control byte written (#401)
Player pets also trigger this nameplate hook function. These pet names
are already translated as they are directly in the game files. Pets
technically do have nameplates, but they are invisible. What happened
here is that when a player pet name hits this function, we write over it
with our control byte (\x04); this is what we use in front of names to
make them not show up as GM players. As the pet names file is loaded
into memory and referenced live, this is adding an extra byte to the
front of the name and shoving the offset of the file over by one byte.
The more player pets that trigger this hook, the more messed up the file
gets as the file pointer looks for null terminated strings. It just ends
up being a bunch of our invisible control characters.

This checks to make sure the text is actually Japanese before attempting
to write the name.