fix: more nameplate problems (#402)
Since we don't have a way to identify an NPC from a player, we are
treating them both with the same logic of truncating their names. For
NPCs, we have tons of long names that were being cut off.

This function looks like it only:

- triggers for players and NPCs
- doesn't trigger on other entities like housing items, player pets,
etc.
- doesn't seem to care how long we write the name

As I still don't know how to identify whether an entity is a player or
an NPC, we're just treating both of these as the same and adding this
"control byte" (\x04) to every name. It doesn't appear to impact NPC
names at present, but I'll let the players test this for me. :) We also
now don't seem to need to truncate the names anymore (outside of what
transliterate_player_name is already truncating).

This also refactors transliterate_player_name to be significantly more
performant by using constants and using an infinite lru_cache for the
session of the game.