feat: port hooks from pymem -> frida. (#404)
This is a significant refactor to move from creating hooks manually with
pymem and asm and instead, use a much more mature instrumentation
library called "Frida" to do hooking. Pymem has been great for remote
editing of memory in a process, but there's a lot of built-in
functionality that has been missing that I have been building myself
over the years.

On top of that, Pymem is having issues getting calls like
PyRun_SimpleString working internally within a process on later versions
of Python, which this program does. This future proofs dqxclarity to
work on later versions of Python, so we can eventually migrate to 3.14
32-bit (from 3.11). Additionally, we didn't have unhooking completely
working, but Frida has it very easily built in, which makes development
significantly easier.

Another side plus is that the way Frida hooks and executes code appears
to work on Linux/Steam deck again, so this will make several users
happy.

All in all, this is a significant change and needs some testing before
cutting a release.